change types to double

+++ backend/src/main/java/org/codecool/fitnesstracker/fitnesstracker/service/analytics/DailyAnalytics.java
@@ -32,7 +32,7 @@ public class DailyAnalytics implements AnalyticDuration {
 
     @Override
     public List<AnalyticDailyDTO> getAnalytics(int userBaseLineCalorieRequirement, User user) {
-        int sumDailyCalorie = calorieService.getCalorieFromDate(getStartingDate(), user).stream().mapToInt(cal -> cal.calorie()).sum();
+        double sumDailyCalorie = calorieService.getCalorieFromDate(getStartingDate(), user).stream().mapToDouble(cal -> cal.calorie()).sum();
         int sumDailyActivity = activityService.getActivityFromDate(getStartingDate(), user).stream().mapToInt(cal -> cal.calorie()).sum();
         List<AnalyticDailyDTO> analyticDailyDTOS = new ArrayList<>();
